update: Notes/NoteList
