+++ src/pages/Notes/Loadable.js
@@ -1,0 +1,3 @@
+import loadable from '../../utils/loadable';
+
+export default loadable(() => import('./index'));
